+++ src/lib.rs
@@ -9,5 +9,8 @@ mod settings;
 /// Localization module
 mod localization;
 
+/// Keycode module
+mod keycode;
+
 /// Application module re-export
 pub use app::ClickStormApp;